Move compoenents, folders, assets, utils, types folders to 'src' folder
