+++ app/Root_layout.tsx
@@ -7,19 +7,19 @@ import { createBottomTabNavigator } from '@react-navigation/bottom-tabs';
 import { Image } from 'react-native';
 import * as Notifications from 'expo-notifications';
 // Function imports
-import { sendPushNotification, registerForPushNotificationsAsync } from '../utils/NotificationHandler';
+import { sendPushNotification, registerForPushNotificationsAsync } from '../src/utils/NotificationHandler';
 // component imports
 // TODO: Find a way to simplify this -- this is a lot of seperate line imports
-import HomeScreen from '../components/screens/HomeScreen';
-import GameScreen from '../components/screens/GameScreen';
-import MessageListScreen from '../components/screens/MessageListScreen';
-import MessageViewScreen from '../components/screens/MessageViewScreen';
-import DebugScreen from '../components/screens/DebugScreen';
-import PlaceholderScreen from '../components/screens/PlaceholderScreen';
-import CalendarTestScreen from '../components/screens/CalendarTestScreen';
-import WebviewScreen from '../components/screens/WebViewScreen';
+import HomeScreen from '../src/screens/HomeScreen';
+import GameScreen from '../src/screens/GameScreen';
+import MessageListScreen from '../src/screens/MessageListScreen';
+import MessageViewScreen from '../src/screens/MessageViewScreen';
+import DebugScreen from '../src/screens/DebugScreen';
+import PlaceholderScreen from '../src/screens/PlaceholderScreen';
+import CalendarTestScreen from '../src/screens/CalendarTestScreen';
+import WebviewScreen from '../src/screens/WebViewScreen';
 
-import { NotificationTokenContext } from '../utils/NotificationHandler';
+import { NotificationTokenContext } from '../src/utils/NotificationHandler';
 
 Notifications.setNotificationHandler({
   handleNotification: async () => ({
@@ -58,19 +58,19 @@ const ROUTE_NAMES = {
 const getIcon = (routeName: string): ImageSourcePropType => {
   switch (routeName) {
     case ROUTE_NAMES.Home:
-      return require("./assets/navigation-tab-icons/menu-icon-home.png");
+      return require("../assets/navigation-tab-icons/menu-icon-home.png");
     case ROUTE_NAMES.Messages:
-      return require("./assets/navigation-tab-icons/menu-icon-messages.png");
+      return require("../assets/navigation-tab-icons/menu-icon-messages.png");
     case ROUTE_NAMES.Insights:
-      return require("./assets/navigation-tab-icons/menu-icon-insights.png");
+      return require("../assets/navigation-tab-icons/menu-icon-insights.png");
     case ROUTE_NAMES.Game:
-      return require("./assets/navigation-tab-icons/menu-icon-game.png");
+      return require("../assets/navigation-tab-icons/menu-icon-game.png");
     case ROUTE_NAMES.Debug:
-      return require("./assets/navigation-tab-icons/menu-icon-debug.png");
+      return require("../assets/navigation-tab-icons/menu-icon-debug.png");
     default:
-      return require("./assets/navigation-tab-icons/menu-icon-error.png");
+      return require("../assets/navigation-tab-icons/menu-icon-error.png");
   }
-};
+}
 
 function MainStackNavigator() {
   return (
@@ -123,7 +123,7 @@ function MainTabNavigator() {
   );
 }
 
-export default function App() {
+export default function RootLayout() {
   
   const [expoPushToken, setExpoPushToken] = useState('');
   const [notification, setNotification] = useState<Notifications.Notification | undefined>(
