+++ src/assets/navigation-tab-icons/menu-icon-debug.png
