+++ src/assets/navigation-tab-icons/menu-icon-game.png
