+++ src/assets/navigation-tab-icons/menu-icon-home.png
