+++ src/assets/navigation-tab-icons/menu-icon-insights.png
