+++ src/assets/placeholders/debug-penguin-helmet.png
