+++ src/components/game_components/GuessWarningPopup.tsx
