+++ src/components/game_components/Letter.tsx
