+++ src/components/game_components/PuzzleView.tsx
