+++ src/screens/GameScreen.tsx
@@ -4,8 +4,8 @@
 
 import React from 'react';
 import { View, StyleSheet } from 'react-native';
-import Game from '../game_components/Game'; // Adjust the import path if needed
-import { GameState, NewGameState } from '../game_components/Game';
+import Game from '../components/game_components/Game'; // Adjust the import path if needed
+import { GameState, NewGameState } from '../components/game_components/Game';
 
 const GameScreen = ({ navigation}) => {
     // This is a set of puzzles for testing purposes.
