+++ src/types/message.type.ts
